backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.2.0
langchain-community==0.2.0
chromadb==0.5.0
sentence-transformers==3.0.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.3.0
groq==0.11.0
python-dotenv==1.0.1
httpx==0.27.0